% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mojito-package.R
\docType{data}
\name{azn}
\alias{azn}
\title{AstraZeneca London Stock Exchange share price}
\format{A \code{data.frame}:
\describe{
  \item{date}{The date.}
  \item{open,high,low,close}{The opening price, high, low and closing price.}
  \item{volume}{Volume of shares traded}
}}
\source{
Various financial websites carry the data. Search for "UK share prices"
and you'll find some.
}
\description{
Daily prices for LON: AZN
}
\keyword{data}
