% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mojito-package.R
\docType{data}
\name{dv}
\alias{dv}
\title{Death Valley maximum daily temperatures}
\format{
A \code{data.frame} with dates, maximum daily temperature, and the
  expected values and residuals from a generalized additive mixed model
  allowing for an annual cycle.
}
\description{
Daily maximum temperatures recorded in Furnace Creek, Death Valley. Provided
on request by National Centers for Environmental Information:
https://www.ncdc.noaa.gov/
}
\keyword{data}
