% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mojito-package.R
\docType{data}
\name{heatwave.dat}
\alias{heatwave.dat}
\title{Maximum daily temperatures}
\format{
A \code{vector} of numeric values of temperatures.
}
\source{
H. C. Winter and J. A. Tawn "Modelling heatwaves in central France: a case-study in extremal dependence", Appl. Statist., 65 (2016), 345 -- 365. EUROPEAN CLIMATE ASSESSMENT & DATASET (ECA&D), THESE DATA CAN BE USED FREELY PROVIDED THAT THE FOLLOWING SOURCE IS ACKNOWLEDGED: 	Klein Tank, A.M.G. and Coauthors, 2002. Daily dataset of 20th-century surface air temperature and precipitation series for the European Climate Assessment. Int. J. of Climatol., 22, 1441-1453. Data and metadata available at http://www.ecad.eu
}
\description{
Maximum daily temperatures in deg C during summer months only.  Data has been pre-processed from the published data.
}
