% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play_mojito_game.R
\name{play_mojito_game}
\alias{play_mojito_game}
\title{Mojito Game}
\usage{
play_mojito_game()
}
\description{
This function is just use to launch the shiny app we created that
 let you play a game called : The mojito game. There is no argument to pass 
 to this function since it just launch the game.
}
\examples{
play_mojito_game()
}
\author{
ggplot3
}
