% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mojito-package.R
\docType{data}
\name{venezuelaRain}
\alias{venezuelaRain}
\title{Venezuela rainfall data}
\format{A \code{data.frame}:
\describe{
  \item{Year,Month,Day}{Numeric values for year, month and day.}
  \item{Rainfall}{Daily rainfall.}
}}
\description{
Daily rainfall data in Venezuela from 1961 to 1999.
}
\keyword{data}
